+++ inc/lib/init.php
@@ -17,8 +17,8 @@ namespace Tofino\Init;
  */
 function php_version_check() {
   $php_version = phpversion();
-  if (version_compare($php_version, '5.5.9', '<')) {
-    wp_die('<div class="error notice"><p>' . __('PHP version >= 7.2.0 is required for this theme to work correctly.', 'tofino') . '</p></div>', 'An error occured.');
+  if (version_compare($php_version, '7.3.0', '<')) {
+    wp_die('<div class="error notice"><p>' . __('PHP version >= 7.3.0 is required for this theme to work correctly.', 'tofino') . '</p></div>', 'An error occured.');
   }
 }
 add_action('after_setup_theme', __NAMESPACE__ . '\\php_version_check');
@@ -106,3 +106,17 @@ function add_post_name_body_class(array $classes) {
   return $classes;
 }
 add_filter('body_class', __NAMESPACE__ . '\\add_post_name_body_class');
+
+
+/**
+ * Add Google Tag Manager function call which is 
+ * supposed to be placed after opening body tag.
+ *
+ * @since 3.3.0
+ */
+function add_custom_body_open_code() {
+  if (function_exists('gtm4wp_the_gtm_tag')) {
+    gtm4wp_the_gtm_tag();
+  } 
+}
+add_action('wp_body_open', __NAMESPACE__ . '\\add_custom_body_open_code');